requirements.txt
click
mypy
pillow
pyocr
regex
